Move tasks function into tiling module
